feat: initial logic
A sidecar container that runs alongside gosmee client. gosmee forwards
events to the sidecar, which runs logic to verify that the client is
alive.

The sidecar is writing heartbeats to a shared volume for each time it
verifies the client is alive.

The liveness probe on the client is verifying receiving heartbeats
and restart the client if they stop arriving.

Signed-off-by: Yftach Herzog <yherzog@redhat.com>